+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -8,8 +8,8 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.Reservation;
-import roomescape.ReservationTime;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationTime;
 
 @Repository
 public class ReservationDao {
