+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.ReservationTime;
+import roomescape.entity.ReservationTime;
 
 @Repository
 public class ReservationTimeDao {
