+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.dto;
 
 import java.time.LocalDate;
-import roomescape.Reservation;
+import roomescape.entity.Reservation;
 
 public record ReservationResponse(long id, String name, LocalDate date, ReservationTimeResponse time) {
